DM-23024: Allow multiple datasets to be associated with a single file on ingest

+++ python/lsst/daf/butler/datastores/posixDatastore.py
@@ -290,13 +290,16 @@ class PosixDatastore(FileLikeDatastore):
             Attempt to remove a dataset that does not exist.
         """
         # Get file metadata and internal metadata
-        location, storefFileInfo = self._get_dataset_location_info(ref)
+        location, _ = self._get_dataset_location_info(ref)
         if location is None:
             raise FileNotFoundError(f"Requested dataset ({ref}) does not exist")
 
         if not os.path.exists(location.path):
             raise FileNotFoundError(f"No such file: {location.uri}")
-        os.remove(location.path)
+
+        if self._can_remove_dataset_artifact(ref):
+            # Only reference to this path so we can remove it
+            os.remove(location.path)
 
         # Remove rows from registries
         self._remove_from_registry(ref)
@@ -340,7 +343,7 @@ class PosixDatastore(FileLikeDatastore):
                 raise FileNotFoundError(f"Could not retrieve Dataset {ref}.")
             if transfer is None:
                 # TODO: do we also need to return the readStorageClass somehow?
-                yield FileDataset(ref=ref, path=location.pathInStore, formatter=storedFileInfo.formatter)
+                yield FileDataset(refs=[ref], path=location.pathInStore, formatter=storedFileInfo.formatter)
             else:
                 # TODO: add support for other transfer modes.  If we support
                 # moving, this method should become transactional.
